Update mcmodel.py
Went through and made the reccommended changes.  Context is now passed as a function parameter for locate_image and read_model.  Added commets that describe how the uv coords work.  I think before I was a bit confused of the concept of docstrings but I believe I better understand it and hopefully implemented it correctly.