Transformed entity spawning to match closer to mob
Properly proxies full imports of rigs, which ensures drivers still work and no deparenting artifact. No longer have the option to 'make real' or not, must be spawned as an appended item. Option in the future to support linking, but likely not helpful. Added initial, non-functional UI for block spawning, and generally updated line and spacings to be more consistent with python style (much of the diff coming from this). More functions made reusable in the spwan_util file for both entities and mob spawning.