Better identify when to prompt resetting resoruce pack paths
Similar to prior commit, this change ensures that the popup is made whenever trying to spawn an effect/item/model where we know up front the filepath doesn't exist. Often happens if you are downloading a file save on another machine with absolute paths. This change prompts the user to take the action to reset paths to defaults so that they have an immediate way to resolve the issue, but can click off if they want to manually fix the paths in advanced settings.