Added backup for MTLs when editing their contents MTL files need to be edited in order for OBJ importing to work properly when using a non-standard colorspace like ACES or AgX. Of course, editing MTL files and then reverting the edits can be error prone, and if done wrong the user won't be able to retrieve the original MTL.
This can be fixed by creating a new folder called "ORIGINAL_MTLS" with
pathlib, and then copying the MTL to that new folder before editing. Now
that the original is backed up, we don't need to revert edits.

Of course the editing section will still run regardless of there's edits
or not, so we'll need to look into checking if an edit was made (perhaps
by comparing timestamps).