Moving to use more generic texture importing for models.
Reusing the general purpose operator for generating materials. It does
not expose the setting values for all of the ways to load the material,
we can think more about that in a future release where there is texture
pack stacking and potentially scene-level settings for the prep settings
(as opposed to forcing a popup per each use ahead of time).