Replaced arguments for matprep_cycles Originally matprep_cycles used to take multiple arguments, each one representing a seperate option. This was annoying as it meant the function has an unfathomable amount of arguments to keep track of.
This has been replaced with a dataclass (called PrepOptions) that keeps
track of all the options selected by the user. Not only does it make it
slightly easier for developers, it also gives the following benefits:
	- More info on what types the different options are
	- Better organization

Previously we couldn't do this due to 2.7x support, hence this being in
mcprep-3_5-recore.