Forcing users to restart if there is any update detected.
This applies to both manual as well as addon-updater updates. The reason
for this change is to avoid errors that will happen in future versions
when we rename modules, remove functions, and generally revamp the
structure of MCprep. Many users encounter errors due to not restarting
Blender already, even with 'typical' updates.