Added Poetry pyproject.toml Why? Poetry means we can easily define Python version, dependencies, etc. without having to force developers to do everything by hand (provided they want to use Poetry, it's fine if they don't c:).
The reason I'm doing it now is due to an upgrade on Arch Linux from
Python 3.10 to Python 3.11, breaking all virtual environments. Poetry is
useful especially on Linux where using the system Python tends to be an
annoyence.

The options I chose are based on MCprep 3.5, which mean the following:
- Blender 2.8 bpy
- Python 3.7 (as that's what 2.8 uses)

These are not set in stone, but I think those are reasonable for MCprep