Fixed Shellcheck errors in compile.sh The compile scripts had some small errors (like a missing exit after cd) but also 2 big errors: SC2308 and SC2115
https://www.shellcheck.net/wiki/SC2308
https://www.shellcheck.net/wiki/SC2115

SC2308 was due to the use of `expr substr` when checking to see if the
OS in use was Linux. According to the shellcheck wiki, this is not
portable and as such shouldn't be used. This was replaced with a pipe
expression that passes to `cut -c 1-5`

SC2115 is due to an `rm -rf "$i/$NAME/" statement, which if the variable
is not set or null, a person's system can be massively messed up. Here's
a case where due to this exact issue, the Steam client ended up deleting
everything on the root directory owned by the user:
https://github.com/ValveSoftware/steam-for-linux/issues/3671

While it's great that these issues were caught ahead of time, these
should have never been in MCprep's build scripts. We already know that
MCprep's build scripts can create folders named ~, causing their own
headaches (like I dealt with a while back). We seriously need to start
enforcing shellcheck for MCprep's build scripts.