Replaced compile scripts with mcprep-build (Saleman pitch time) mcprep-build is our new, state-of-the-art build system custom made for the MCprep project that's cross platform! Now we get less headaches from dealing with Bash and the Windows shell as this new build system is written in Python!
Jokes aside, this new build system is much more maintainable then the
old build system. I still need to add autopopulation, but it's 99% the
same, minus fast mode.

In the future I want to turn this into a general purpose
bpy-addon-builder sorta thing, but let's not get ahead of ourselves.