Added some basic code for i18n loading
Pop vs Soda

Now MCprep has the ability to load JSON files for i18n support. It
doesn't actually do anything with the data right now, but basic support
is there.

This method of loading also handles subdirectories, which means that we
could have folders for seperate dialects under a larger folders for the
language as a whole, such as Pop English and Soda English in a larger
English folder, or a Chinese folder with different dialects of Chinese.

Overall this is a big moment for MCprep, and paves the way to allow us
to fulfill the goal of "Let no Minecraft animator be confused by
MCprep's UI simply because of the language they speak"