Establish new cross platform test suite using unittest library.
This sets up the framework and ability to test the addon on multiple
versions of blender. It continues to use the blender_execs.txt file
to indicate which versions of blender are for running tests, but expands
on controls to let the user specify which specific versions of blender
they want to test, and with nice commandline documentation for args.