Changed behavior of skin swap to be more nuanced
Skin swap originally would change all image textures on all materials of
the selected object, regardless of whether the user wanted to or not.
This is not ideal for rigs that need multiple seperate textures.

To expand the use case of skin swap, skin swap will now use a more
nuanced method of swapping skins based on node name, which is an already
established precedent from the MCprep Cycles Optimizer