Prevent collection-effect obj particle src from being cleaned up.
Before this change, if a collection effect included a particle system
where the source of the particles was an object (as opposed to an entire
collection), then on save and reload of that file, the particle object
would have been cleaned up, as it was not linked to the scene and for
some reason, since forever, blender does not count usage of an object
as a particle emitter as an object use reference when it comes to its
own internal cleanup.

Works now, tested by saving and reloading a file after spawning in a
version of the firework tutorial where I didn't use a collection as the
firework source.