Added BSD license to commonmcobj_parser.py
(copied from commonmcobj_parser.py)

Normally, I wouldn't do dual licensing, but in this case, it makes sense
as it would allow developers to reuse this parser for their own uses
under more permissive terms. This doesn't change anything related to
MCprep, which is GPL, as BSD 3-Clause is compatible with GPL. The only
part that might conflict is Clause 3, but it could be argued that one
can't do that under GPL anyway, or any license for that matter, and that
Clause 3 is just a reminder to developers.