feat: Upstream Vivy material system
Vivy is a new material system that is based around the concept of
templates instead of generation. I created it in my spare time
in 2023 and used it for my new year's piece

https://www.standingpad.org/posts/2023/12/new-years/#vivy

Initially, I have said that I didn't want to upstream Vivy in MCprep to
avoid users from opening bug reports after underestimating the
complexity of using the system. That said, #274 already had plans to
revamp the material system for greater user flexibility, and I'm unable
to maintain Vivy on my own (given the fast pace of upstream MCprep).
Therefore, I've decided to begin the work of upstreaming Vivy into
MCprep itself.

I initially tried to merge Vivy into a separate branch, but that turned
out to be impossible due to massive changes in mainline MCprep. Thus,
I've included the Vivy files here and added the necessary changes to
`conf.py` and `load_modules.py`. This was made easier by the fact that
Vivy does not change anything internal, mostly due to laziness on my
part (instead, I just copied and pasted functions I needed to modify for
Vivy). As such, everything in Vivy is a separate operator, using the
`vivy` prefix instead of `mcprep`. I've not included the necessary
reference changes for Vivy, since I don't want to start overriding
operators yet.

As it stands, Vivy is lacking in the following areas:
- UI for setting up materials
- Documentation for using
- Integration in the rest of MCprep (like mob rigs)

In addition, Vivy also has the following issues due to the way its
designed:
- Use of `VIVY` and `vivy` prefixes
- Lose integration (again, due to laziness on my part)
- No tests

Seeing as this is more of a long term project however, I think we can
get these tackled sometime by MCprep 4.0. Once those are done, we'd also
need to discuss how this would coexist with the existing generation
based system (should we choose to keep it).

Closes #274

Signed-off-by: Mahid Sheikh <mahid@standingpad.org>