feat: Mark Vivy as an experimental feature
This makes it easier for users to opt-in Vivy if they wish to do so.
This also makes Vivy officially the first experimental feature in
MCprep.

This requires 2 levels of opt-in: first enabling the experimental
feature set, and then enabling the Vivy system. This is due to all of
the overrides we do in the UI for Vivy.

Signed-off-by: Mahid Sheikh <mahid@standingpad.org>