feat: Add ability for users to locally refresh MCprep Resources
Sometimes, there are circumstances where a user may want to update the
MCprep resources folder locally without having to wait for an MCprep
update. Although we plan to eventually implement separate updates for
assets, this adds the ability to locally refresh textures based on the
user's locally installed version of Minecraft.

In theory this can also be used as part of the move towards not bundling
Minecraft textures, which A. are a legal gray area and B. could be a GPL
violation (at least in discussions I had with the FSF, their logic being
that Minecraft is a proprietary game, though I'd argue this is more of a
gray area since we're dealing with non-code assets).

This also updates push_latest.sh to use the now moved script

Signed-off-by: Mahid Sheikh <mahid@standingpad.org>