refactor: Move custom node properties to PropertyGroup
Closes #668

Previously, for the last 7 years, MCprep added custom properties to
nodes directly. However, in Blender 5.0, this was changed (see
https://projects.blender.org/blender/blender/commit/7276b2009a8819619263a1b897389662307b0ee0),
causing MCprep to break when using the `generate` function.

To fix this, this patch moves those custom properties to a `PropertyGroup`
called `MCprepNodeProps`, which is then registered and added as part of
`bpy.types.Node`. Apparently this is the intended way for custom
properties to be added to nodes, so it should work in previous versions
of Blender.

As part of backwards compatibility with old files, the
`np_is_mcprep_node_prop` function was added to `util`, which performs 3
steps:

- Check if the old properties exist on the node
- If old property exists, set the new equivilent
- Return new equivilent of property

When ran, this function will act as both a compatibility layer and
migration function.

Credit goes to [snzhongcheng](https://github.com/snzhongcheng) on the
MCprep Discord server for suggesting this solution, and
[BradyAJonston](https://github.com/BradyAJohnston) from the Erindale
server for confirming this is the intended way of adding custom properties
to nodes (and thus shouldn't break in future releases of Blender)

Signed-off-by: Mahid Sheikh <mahid@standingpad.org>