refactor: Create Vivy V2 JSON types
This creates the basic types for Vivy V2, which cleans up the overall
JSON structure. This does not yet implement support for V2 in MCprep,
but it sets the foundation for later.

Also included in this change is the addition of typing_extensions 4.7.1,
which backports some new typing features to older versions of Python
(this comes directly from the Python Foundation). We're using an old
version here due to Blender 2.83 using Python 3.7.

Signed-off-by: Mahid Sheikh <mahid@standingpad.org>