feature: Add Import Multiple OBJs as Chunks
This adds the "Import Multiple OBJs	as Chunks" feature, which allows
users to import multiple OBJs as part of a larger chunk, and have the
offsets automatically handled by MCprep.

This is done by first loading the CommonMCOBJ headers of all the
selected OBJs, making sure the block scales and world are the same.
Then, we go through each header and store both the min and max export
bounds for X, Y, and Z in 3 arrays. Then, we sort those 3 arrays, and
create a new min bound using the first values of all 3 arrays, and a new
max value using the last values in all 3 arrays. Then, we calculate a
center point and negate it. Finally, we import all of the OBJs, undo the
export offset (and block offset), and apply the new offset to them based
on the center point.

The nice thing about this method is that it doesn't matter if there's
holes in the final assembled world, nor does it matter if it's a perfect
cube or if there's protruding bits.

Signed-off-by: Mahid Sheikh <mahid@standingpad.org>