fix: Add backface culling to redstone torches
It seems that redstone torches repeaters, and comparators need
special handling in newer versions of Minecraft when it comes to
rendering. Specifically, a backface culling setup needs to be
added to the material that subtracts the backface from the alpha 
channel of the texture, making backfaces transparent (and thus
properly rendering redstone torches).

This could in theory also be needed for `Non-solid` blocks, but based on
the list provided by Minecraft.Wiki (https://minecraft.wiki/w/Opacity),
it seems redstone torches, repeaters, and comparators are the only thing
that require this special handling (perhaps due to that "second layer"
that they have as of 1.21.5?). On the top of my head, everything else in
that list should render fine without backface culling.

Unfortunately this does impact the old redstone torch texture, so extra
handling will need to be done for that. Now I wish CommonMCOBJ V1 had MC
version in the header, that would make things so much simpler -_-

Fixes #647

Signed-off-by: Mahid Sheikh <mahid@standingpad.org>