feat: Add drag and drop handler for JSON models
This is a small thing, but Blender 4.1 introduced the new FileHandler
API, which allows defining custom drag and drop behavior. Since we
already have an operator for importing JSON models, we can extend it
relatively easily by creating a subclass of `FileHandler`.

Since it's exclusive to 4.1 and above, we wrap the `FileHandler` type
around another type called `FileHandlerIfSupported`, which otherwise
will simply be a dummy type at runtime if the Blender version is 4.0 or
below. In addition, we also only register the
`MCPREP_FH_import_minecraft_model_file` class if Blender 4.1 and above
is being used.

Signed-off-by: Mahid Sheikh <mahid@standingpad.org>