refactor: Deduplicate shared OBJ importer functionality
To avoid large chunks of duplicated code between 2 OBJ import operators,
this PR extracts the common parts into 2 functions.

One function, `validate_and_return_header` performs file validation
(making sure the file exists, is a file, is an OBJ and/or corrects MTL
selection, etc) and extracts the header (if it exists). Another,
`import_obj_file`, performs the import steps and object property setups,
as well as perform offsets if provided.

A downside of the duplication is that warnings are a bit harder to
propogate. For now, we've simply done a sort of "hack" where we have a
dummy exception type for warnings, and use that for anything intended to
be shown in an `INFO` report (as opposed to an `ERROR` report). Maybe it
would be better to create a deticated object (perhaps `MCprepInfo`), but
that's beyond the scope of this commit.

Signed-off-by: Mahid Sheikh <mahid@standingpad.org>