fix: Add `colormap/grass` back in JSON definitions
This reverts a small change made in #655 in which `colormap/grass` was
altered to `short_grass`. While this works well for grass in new
exports, this causes issues in older exports, like what we have in our
test suite. Thus, we need to add it back into the JSON file to handle
older exports.

Signed-off-by: Mahid Sheikh <mahid@standingpad.org>