feat: Set shading to flat for Mineways imports
Mineways OBJs seem to have weird artifacts caused by vertex normals
being set (and thus in Blender being shaded smooth). While this is being
tracked in erich666/Mineways (#151), a workaround is needed for existing
OBJs to reduce issues.

This is done by simply setting the `use_smooth` property to `False` on
the active selection, which will be the OBJ after importing, only if the
OBJ is a Mineways export. This behavior can be disabled by calling the
function with `mineways_fix_smooth_shading_artifacts` set to `False`. As
this is done through the `foreach_set` function, there shouldn't be much
of a hit to import performance. However, just in case, the end user can
also disable the workaround.

If future releases of Mineways resolve the smooth shading issue, we
could encapsulate it with a header check, either for the version (if
future versions of Mineways completely eliminate setting vertex normals)
or for an export option (if setting vertex normals can be disabled in
the future, which is a more likely outcome).

Signed-off-by: Maryam Sheikh <mahid@standingpad.org>