feat: Add centering options for single OBJ imports
This is mostly for parity with importing OBJs as chunks. Like chunk
imports, this will only be performed on OBJs that export CommonMCOBJ
metadata.

In addition, this also adds a warning for OBJs that don't use the
CommonMCOBJ header. Given it's been almost two years, I think it's safe
to begin warning users about old OBJs, since most at this point should
be exported with newer exporters (usually for newer Minecraft updates).

Signed-off-by: Maryam Sheikh <mahid@standingpad.org>